+++ src/main/java/com/ar/dev/tierra/api/controller/FacturaController.java
@@ -13,7 +13,7 @@ import com.ar.dev.tierra.api.model.Factura;
 import com.ar.dev.tierra.api.model.JsonResponse;
 import com.ar.dev.tierra.api.model.Usuarios;
 import java.io.Serializable;
-import java.math.BigInteger;
+import java.math.BigDecimal;
 import java.util.Date;
 import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -65,7 +65,7 @@ public class FacturaController implements Serializable {
         factura.setUsuarioCreacion(user.getIdUsuario());
         factura.setFechaCreacion(new Date());
         factura.setEstado("INICIADO");
-        factura.setTotal(BigInteger.ZERO);
+        factura.setTotal(BigDecimal.ZERO);
         int idFactura = facturaDAO.add(factura);
         JsonResponse msg = new JsonResponse("Success", String.valueOf(idFactura));
         return new ResponseEntity<>(msg, HttpStatus.OK);
